requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.2.0
matplotlib>=3.4.2
diffusers>=0.14.0
transformers>=4.11.0
